+++ example/string-literal/string-literal.ts
@@ -1,7 +1,7 @@
 import { Lexer } from "../../src";
 
 export const lexer = new Lexer.Builder()
-  .ignore(Lexer.whitespaces)
+  .ignore(Lexer.whitespaces())
   .define({
     simpleStr: Lexer.stringLiteral(`'`).or(Lexer.stringLiteral(`"`)),
     multilineStr: Lexer.stringLiteral("`", { multiline: true }),